backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
google-auth==2.40.2
google-auth-oauthlib==1.2.4
google-api-python-client==2.172.0
cloudinary==1.44.1
requests==2.32.3
pydantic==2.10.6
